# Improve useRestyle performance