--- a/src/composeRestyleFunctions.ts
+++ b/src/composeRestyleFunctions.ts
@@ -5,6 +5,7 @@ import {
   BaseTheme,
   Dimensions,
   RNStyle,
+  RestyleFunction,
 } from './types';
 import {AllProps} from './restyleFunctions';
 
@@ -26,18 +27,19 @@ const composeRestyleFunctions = <
   const properties = flattenedRestyleFunctions.map(styleFunc => {
     return styleFunc.property;
   });
-  const funcs = flattenedRestyleFunctions
-    .sort(
-      (styleFuncA, styleFuncB) =>
-        Number(styleFuncB.variant) - Number(styleFuncA.variant),
-    )
-    .map(styleFunc => {
-      return styleFunc.func;
-    });
+  const propertiesMap = properties.reduce(
+    (acc, prop) => ({...acc, [prop]: true}),
+    {} as Record<keyof TProps, true>,
+  );
+
+  const funcsMap = flattenedRestyleFunctions.reduce(
+    (acc, each) => ({[each.property]: each.func, ...acc}),
+    {} as Record<keyof TProps, RestyleFunction<TProps, Theme, string>>,
+  );
 
   // TInputProps is a superset of TProps since TProps are only the Restyle Props
-  const buildStyle = <TInputProps extends TProps>(
-    props: TInputProps,
+  const buildStyle = (
+    props: TProps,
     {
       theme,
       dimensions,
@@ -46,15 +48,20 @@ const composeRestyleFunctions = <
       dimensions: Dimensions;
     },
   ): RNStyle => {
-    const styles = funcs.reduce((acc, func) => {
-      return Object.assign(acc, func(props, {theme, dimensions}));
-    }, {});
+    const styles = Object.keys(props).reduce(
+      (styleObj, propKey) => ({
+        ...styleObj,
+        ...funcsMap[propKey as keyof TProps](props, {theme, dimensions}),
+      }),
+      {},
+    );
     const {stylesheet} = StyleSheet.create({stylesheet: styles});
     return stylesheet;
   };
   return {
     buildStyle,
     properties,
+    propertiesMap,
   };
 };