--- a/src/createRestyleComponent.tsx
+++ b/src/createRestyleComponent.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import {View} from 'react-native';
 
+import composeRestyleFunctions from './composeRestyleFunctions';
 import {BaseTheme, RestyleFunctionContainer} from './types';
 import useRestyle from './hooks/useRestyle';
 
@@ -13,8 +14,10 @@ const createRestyleComponent = <
     | RestyleFunctionContainer<Props, Theme>[])[],
   BaseComponent: React.ComponentType<any> = View,
 ) => {
+  const composedRestyleFunction = composeRestyleFunctions(restyleFunctions);
+
   const RestyleComponent = React.forwardRef((props: Props, ref) => {
-    const passedProps = useRestyle(restyleFunctions, props);
+    const passedProps = useRestyle(composedRestyleFunction, props);
     return <BaseComponent ref={ref} {...passedProps} />;
   });
   type RestyleComponentType = typeof RestyleComponent;